refactor(components): [date-picker] basic cell (#7924)
- Refactor `h` function to `jsx`.
- Extract props out of `basic-cell-render`
