+++ packages/components/date-picker/src/props/basic-cell.ts
@@ -1,0 +1,12 @@
+import { buildProps, definePropType } from '@element-plus/utils'
+
+import type { ExtractPropTypes } from 'vue'
+import type { DateCell } from '../date-picker.type'
+
+export const basicCellProps = buildProps({
+  cell: {
+    type: definePropType<DateCell>(Object),
+  },
+} as const)
+
+export type BasicCellProps = ExtractPropTypes<typeof basicCellProps>
